requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
openai==1.12.0
pandas==2.1.1
numpy==1.26.0
tqdm==4.66.1
scikit-learn==1.3.0
flask==2.3.3
flask-cors==4.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
pymongo==4.12.0
python-jose==3.4.0
PyJWT==2.8.0
pinecone==6.0.2
httpx==0.27.0
websocket-client==1.7.0
bcrypt==4.0.1
